MAINT: move tera into tools & remove unused code
